requirements.txt
streamlit==1.38.0
requests==2.32.3
plotly==5.22.0
pandas==2.2.2
pillow==10.4.0
openai==1.37.0
supabase==2.5.1
python-dotenv==1.0.1